backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pgvector==0.2.5
pydantic[email]==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
cryptography==42.0.7
celery==5.4.0
redis==5.0.4
flower==2.0.1
google-generativeai==0.5.4
pypdf2==3.0.1
python-docx==1.1.2
openpyxl==3.1.2
httpx==0.27.0
tenacity==8.3.0
